requirements.txt
smartapi-python
pyotp
pandas
python-dotenv
requests
